Fix inconsistent state after discard
the _clear_txstate() method is called from discard without waiting for
the 'DISCARD' command to complete on the server side. This results in a
brief period where the client in in an inconsistent state: the
transaction has been cleared on the client and the connection has been
returned to the connection pool, but the transaction is still open on
the server.

If another client grabs the connection during this period and tries to
start its own transaction during this period, it creates a race
condition between the DISCARD and the MULTI. If the MULTI is executed
first on the server, it causes an '(error) ERR MULTI calls can not be
nested' response error.

Instead, make _clear_txstate a callback for the discard method,
guaranteeing that the connection is not returned to the queue until the
transaction is well and truly discarded. Make sure to pass the response
through to ensure backward compatibility.